pyproject.toml
[project]
name = "aeon"
version = "1.0.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10, <3.13"
dependencies = [
    "colorama (>=0.4.6,<0.5.0)",
    "matplotlib (==3.8.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "pyro-api (>=0.1.2,<0.2.0)",
    "pyro-ppl (>=1.9.1,<2.0.0)",
    "pysam (>=0.22.1,<0.23.0)",
    "torch (>=2.5.1,<3.0.0)",
    "mypy (>=1.14.1,<2.0.0)",
    "types-colorama (>=0.4.15.20240311,<0.5.0.0)",
    "pandas-stubs (>=2.2.3.241126,<3.0.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
aeon = "aeon.aeon:main"